LibAssistant-issue/requirements.txt
grpcio
grpcio-tools
protobuf
SQLAlchemy
